Refactored create timesheet Api

+++ src/main/java/com/zaga/model/entity/DailyTimesheet.java
@@ -24,7 +24,7 @@ public class DailyTimesheet extends PanacheMongoEntity {
     public String projectId;
     public String dailyTimesheetId;
     public String projectName;
-    public Float hours;
+    public Double hours;
     public LocalDate date;
     public String supportTicket;
     public List<String> clientOwners;
